% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/history.R
\name{abridged}
\alias{abridged}
\title{Convert a full history list from Hipchat to an abridged data.frame}
\usage{
abridged(history)
}
\arguments{
\item{history}{list. The raw output from the Hipchat API.}
}
\value{
A data.frame with columns \code{c('id', 'from', 'message', 'color', 'time')}
}
\description{
Convert a full history list from Hipchat to an abridged data.frame
}
\seealso{
\code{\link{hipchat_history}}
}

