% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/send.R
\name{determine_method}
\alias{determine_method}
\title{Determine the correct HTTP method to use for a given API route.}
\usage{
determine_method(url)
}
\arguments{
\item{url}{character. The HTTP URL.}
}
\description{
If multiple methods are available (e.g., GET and DELETE), a warning will be issued
and only the first method will be used.
If no methods are available, it is probably an invalid API route.
In this case, a warning will be issued and GET will be used.
}
\examples{
stopifnot(hipchat:::determine_method('nonexistent') == 'GET') # Warns about no method
stopifnot(hipchat:::determine_method('room') == 'GET') # Warns about multiple methods
stopifnot(hipchat:::determine_method('user/some@guy.com/message') == 'POST') # No warning
}

