% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{package}
\name{hipchat}
\alias{hipchat}
\alias{hipchat-package}
\title{Description: Communication with the hipchat API, including sending
    messages to rooms, creating and destroying rooms, changing the topic,
    and manipulating and messaging users.}
\description{
Description: Communication with the hipchat API, including sending
    messages to rooms, creating and destroying rooms, changing the topic,
    and manipulating and messaging users.
}

