% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hipchat_api_token}
\alias{hipchat_api_token}
\title{Hipchat API token.}
\usage{
hipchat_api_token()
}
\value{
the Hipchat API token in the environment variable \code{HIPCHAT_API_TOKEN}
   or the R option \code{getOption('hipchat.api_token')}.
}
\description{
Hipchat API token.
}

