% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/room.R
\name{hipchat_create_room}
\alias{hipchat_create_room}
\title{Create a new Hipchat room. (Must have privileges)}
\usage{
hipchat_create_room(room_name, topic = NULL, guest_access = FALSE,
  owner_user, privacy = "public")
}
\arguments{
\item{room_name}{character. Room name.}

\item{topic}{character. Topic name (optional).}

\item{guest_access}{logical. Whether or not to enable guest access for this room.
By default, \code{FALSE}.}

\item{owner_user}{character or logical. The id, email address, or mention name
(beginning with an @) of the room's owner. Defaults to the current user. (Optional)}

\item{privacy}{character. Whether the room is available for access by other users or not.
Must be either \code{'public'} or \code{'private} (default is \code{'public'}).}
}
\value{
the id of the newly created room.
}
\description{
Create a new Hipchat room. (Must have privileges)
}

