% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/user.R
\name{hipchat_create_user}
\alias{hipchat_create_user}
\title{Create a new Hipchat user.}
\usage{
hipchat_create_user(name, password, email, title, mention_name,
  is_group_admin = FALSE, timezone = "UTC")
}
\arguments{
\item{name}{character. User's full name (1-50 characters). (Required)}

\item{password}{character. User's password. (Required)}

\item{email}{character. User's email. (Required)}

\item{title}{character. User's title. (Optional)}

\item{mention_name}{character. User's mention name. (Optional)}

\item{is_group_admin}{logical. Whether or not this user is an admin. (Optional)
The default is \code{FALSE}.}

\item{timezone}{character. A supported timezone (see note). (Optional)
The default is \code{'UTC'}.}
}
\value{
the id for the newly created user, or NA if creation failed.
}
\description{
Create a new Hipchat user.
}
\note{
for available time zones, see the third column in the table
  given at http://en.wikipedia.org/wiki/List_of_tz_database_time_zones
}
\examples{
\dontrun{
  hipchat_create_user('My name', 'mypassword', 'my.email@org.com',
    'my title', '@MyName')
}
}
\references{
https://www.hipchat.com/docs/apiv2/method/create_user
}

