% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/room.R
\name{hipchat_delete_room}
\alias{hipchat_delete_room}
\title{Delete a Hipchat room.}
\usage{
hipchat_delete_room(room_name_or_id)
}
\arguments{
\item{room_name_or_id}{character or integer.}
}
\description{
Delete a Hipchat room.
}
\examples{
\dontrun{
   hipchat_create_room('Example room')
   hipchat_delete_room('Example room')
}
}

