% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/history.R
\name{hipchat_history}
\alias{hipchat_history}
\title{Show the history for a given Hipchat room.}
\usage{
hipchat_history(room_name_or_id)
}
\arguments{
\item{room_name_or_id}{character or integer.}
}
\description{
Show the history for a given Hipchat room.
}

