% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room.R
\name{hipchat_room_id}
\alias{hipchat_room_id}
\title{Convert a Hipchat room name to an ID.}
\usage{
hipchat_room_id(room_name, api_token = hipchat_api_token())
}
\arguments{
\item{room_name}{character. A character vector giving the room
name. This is vectorized if you would like IDs for multiple rooms.}

\item{api_token}{character. By default, \code{\link{hipchat_api_token}()}.}
}
\value{
A named integer vector, with the names being the room names
  and the values being the room IDs. If a room ID is not found, the
  value returned will be NA, so that if you pass in one room name
  and it does not exist, NA will be returned.
}
\description{
This function will call the \code{/v2/room} API route if necessary to 
fetch the room-id map. If a room has been found before, it will use
the cached value. Otherwise, if it is not found, it will refresh
the cache.
}
\note{
This will call the hipchat API at \code{https://api.hipchat.com/v2/room}
  to fetch the room-id map.
}

