% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/room.R
\name{hipchat_rooms}
\alias{hipchat_rooms}
\title{Get the list of all hipchat rooms.}
\usage{
hipchat_rooms(api_token = hipchat_api_token())
}
\arguments{
\item{api_token}{character. By default, \code{\link{hipchat_api_token}()}.}
}
\value{
an integer vector, with names being room names and values
   being room IDs.
}
\description{
Get the list of all hipchat rooms.
}

