% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{hipchat_send}
\alias{hipchat_send}
\title{Send a message to the hipchat API v2.}
\usage{
hipchat_send(type, var, ..., api_token = hipchat_api_token())
}
\arguments{
\item{type}{character. One of the native Hipchat API domains:
\code{c('addon', 'capabilities', 'emoticon', 'oauth', 'room' 'user')}.}

\item{var}{character. Whatever goes after the \code{type} in the
hipchat API url, typically room ID, name, or user email.}

\item{...}{HTTP arguments to send to the Hipchat API. Unnamed arguments
will be passed in to the URL; named arguments will be assumed to be
API parameters. See the examples section.}

\item{api_token}{character. By default, \code{\link{hipchat_api_token}()}.}
}
\value{
the JSON output response from the Hipchat API.
}
\description{
Send a message to the hipchat API v2.
}
\examples{
\dontrun{
  hipchat_send('room', 'My room', 'notification',
    color = 'green', message = 'Everything looks <b>green</b>',
    notify = FALSE, message_format = 'html')
  # posts to http://
}
}
\references{
https://www.hipchat.com/docs/apiv2
}

