% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hipchat_use_send_error_suppression}
\alias{hipchat_use_send_error_suppression}
\title{Option to handle hipchat_send errors as warnings or errors}
\usage{
hipchat_use_send_error_suppression()
}
\value{
logical in the R option \code{getOption('hipchat.use_send_error_suppression')}.
}
\description{
Option to handle hipchat_send errors as warnings or errors
}

