% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{is.success}
\alias{is.success}
\title{Determine if a status code is a successful status code.}
\usage{
is.success(code)
}
\arguments{
\item{type}{numeric. The HTTP status code}
}
\description{
Determine if a status code is a successful status code.
}

