% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_method}
\alias{match_method}
\title{Find appropriate HTTP method for a given Hipchat URL.}
\usage{
match_method(url)
}
\arguments{
\item{url}{character. The URL to analyze.}
}
\value{
character vector of applicable methods, from
   \code{c("GET", "POST", "PUT", "DELETE"}.
}
\description{
Find appropriate HTTP method for a given Hipchat URL.
}

