% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/hipchat-package.R, R/message.R
\docType{package}
\name{hipchat}
\alias{hipchat}
\alias{hipchat-package}
\title{Description: Communication with the hipchat API, including sending
    messages to rooms, creating and destroying rooms, changing the topic,
    and manipulating and messaging users.}
\usage{
hipchat(room_or_user, message, notify = TRUE, color = "yellow",
  message_format = "text", api_token = hipchat_api_token())
}
\arguments{
\item{room_or_user}{character or integer. The room(s) and/or user(s) to send a
  message to. Room and/or user IDs will also work, but you will first
  have to obtain the correct ID, so it is easier to user the former.
  Rooms can be specified by full name (e.g., \code{"This room"}) and users
  can be specified by email (e.g., \code{"some@user.org"}) or mention
  name (e.g., \code{"@SomeUser"}).

  If IDs are provided, they will be checked against the list of available
  room IDs. If the ID is present, a room ID will be assumed; otherwise,
  a user ID will be assumed.}

\item{message}{character. The message(s) to send. If a message is over
10,000 characters, it will be split up into multiple messages, as
the Hipchat API only allows up to 10,000 characters per message.}

\item{notify}{logical. Whether or not to notify the target recipient.
The default is \code{TRUE}, although \code{FALSE} is less invasive
and useful for non-critical messages.}

\item{color}{character. Available options are: yellow, red, green,
purple, gray, or random. Defaults to yellow. If this is a message
sent to a user, this option will be ignored.}

\item{message_format}{character. Available options are: text or html.
The default is text.}

\item{api_token}{character. Optional API token. The default is
\code{hipchat:::hipchat_api_token()}, which uses
 the environment variable \code{HIPCHAT_API_TOKEN} or the
 R option \code{getOption('hipchat.api_token')} if they are present.}
}
\value{
\code{TRUE} if the message was successfull, or an error otherwise.
}
\description{
Description: Communication with the hipchat API, including sending
    messages to rooms, creating and destroying rooms, changing the topic,
    and manipulating and messaging users.

Send a message to a Hipchat room or user.
}
\examples{
\dontrun{
  hipchat('Some room', 'Hey guys!')
  hipchat('some@user.org', "Hey buddy how's it going?")
  hipchat('Some room', 'Server is down!', color = 'red')
  hipchat('Some room', 'Look at this picture: <img src="https://www.google.com/images/srpr/logo11w.png">', message_format = 'html')

  # A more comprehensive example
  hipchat(c("This room", 'and.this@user.org'), "Really important stuff", notify = TRUE,
    color = 'red', message_format = 'text', api_token = 'my_api_token')
}
}
\seealso{
\code{\link{hipchat_send}}
}

