% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/room.R
\name{hipchat_delete_room}
\alias{hipchat_delete_room}
\title{Delete a Hipchat room.}
\usage{
hipchat_delete_room(room_name_or_id, confirm = TRUE)
}
\arguments{
\item{room_name_or_id}{character or integer.}

\item{confirm}{logical. Whether or not to ask for a confirmation message
before deleting the room. By default, \code{TRUE}. (Deleting rooms
is dangerous!)}
}
\value{
\code{TRUE} or \code{FALSE} according as the room was deleted.
}
\description{
Delete a Hipchat room.
}
\examples{
\dontrun{
   hipchat_create_room('Example room')
   hipchat_delete_room('Example room') # Will ask a confirmation message.
   hipchat_delete_room('Example room', confirm = FALSE) # Dangerous! No confirmation.
}
}

