% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/send.R
\name{hipchat_url}
\alias{hipchat_url}
\title{Hipchat API url.}
\usage{
hipchat_url(..., api_token = hipchat:::hipchat_api_token())
}
\arguments{
\item{...}{concatenate these arguments in to the hipchat URL.}

\item{api_token}{character. By default, \code{\link{hipchat_api_token}()}.}
}
\value{
https://api.hipchat.com/v2/...
}
\description{
Hipchat API url.
}
\examples{
stopifnot(hipchat_url('oauth', 'token', use.auth_token = FALSE) ==
  paste(hipchat:::hipchat_api_url, 'oauth', 'token', sep = '/'))
}

